@@ -11,31 +11,30 @@
   "dependencies": {
     "@headlessui/react": "^1.7.17",
     "@heroicons/react": "^2.0.18",
-    "@onflow/fcl": "^1.3.2",
+    "@onflow/fcl": "^1.9.0",
     "@popperjs/core": "^2.11.8",
-    "@supabase/supabase-js": "^2.38.4",
+    "@supabase/supabase-js": "^2.39.0",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tailwindcss/forms": "^0.5.7",
     "@tanstack/react-table": "^8.10.7",
     "@vercel/analytics": "^1.0.1",
-    "airtable": "^0.12.2",
     "autoprefixer": "^10.4.16",
     "clsx": "^2.0.0",
     "crypto": "^1.0.1",
-    "eslint": "^8.46.0",
-    "eslint-config-next": "^14.0.3",
-    "next": "^14.0.3",
-    "next-auth": "^5.0.0-beta.3",
-    "postcss": "^8.4.27",
+    "eslint": "^8.55.0",
+    "eslint-config-next": "^14.0.4",
+    "next": "^14.0.4",
+    "next-auth": "^5.0.0-beta.4",
+    "postcss": "^8.4.32",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "tailwind-merge": "^2.0.0",
-    "tailwindcss": "^3.3.3",
+    "tailwind-merge": "^2.1.0",
+    "tailwindcss": "^3.3.6",
     "usehooks-ts": "^2.9.1"
   },
   "devDependencies": {
     "@onflow/typedefs": "^1.2.1",
-    "@types/react": "18.2.37",
+    "@types/react": "^18.2.42",
     "encoding": "^0.1.13"
   }
 }
