@@ -1,4 +1,4 @@
-import { ExclamationCircleIcon } from '@heroicons/react/24/outline';
+import { ExclamationCircleIcon } from "@heroicons/react/24/outline";
 
 export default function NotFound() {
   return (
@@ -6,7 +6,9 @@ export default function NotFound() {
       <h2 className="text-3xl tracking-tight font-bold text-slate-900 dark:text-slate-200 flex justify-center">
         <ExclamationCircleIcon className="w-16 h-16" />
       </h2>
-      <p className="mt-4 text-slate-900 dark:text-slate-200">Player with the requested ID could not be found.</p>
+      <p className="mt-4 text-slate-900 dark:text-slate-200">
+        The compare page you are looking for could not be found.
+      </p>
     </div>
   );
 }
