@@ -22,22 +22,22 @@ export default async function BasicInfo({ player }) {
   };
   return (
     <div className="col-span-2">
-      <dl className="divide-y divide-slate-200 dark:divide-slate-700">
+      <dl className="@container/inner divide-y divide-slate-200 dark:divide-slate-700">
         {Object.entries(metadata).map(([key, value]) => (
           <div
             key={key}
-            className="px-2 py-2 grid grid-cols-3 sm:gap-4 sm:px-0"
+            className="px-1 py-1.5 @[16rem]/inner:py-2 grid grid-cols-3 gap-8 @[16rem]/inner:px-0"
           >
-            <dt className="text-sm sm:text-base font-semibold leading-6 text-slate-700 dark:text-slate-400 uppercase">
+            <dt className="text-sm @[16rem]/inner:text-base font-semibold leading-6 text-slate-700 dark:text-slate-400 uppercase">
               {key}
             </dt>
-            <dd className="text-sm sm:text-base leading-6 text-slate-700 dark:text-slate-200 text-right sm:text-left col-span-2 capitalize">
+            <dd className="text-sm @[16rem]/inner:text-base leading-6 text-slate-700 dark:text-slate-200 text-right @[16rem]/inner:text-left col-span-2 capitalize">
               {value}
             </dd>
           </div>
         ))}
-        <div className="px-2 py-2 grid grid-cols-3 sm:gap-4 sm:px-0">
-          <dt className="text-sm sm:text-base font-semibold leading-6 text-slate-700 dark:text-slate-400 uppercase flex items-center space-x-2">
+        <div className="px-1 py-1.5 @[16rem]/inner:py-2 grid grid-cols-3 gap-8 @[16rem]/inner:px-0">
+          <dt className="text-sm @[16rem]/inner:text-base font-semibold leading-6 text-slate-700 dark:text-slate-400 uppercase flex items-center space-x-2">
             <span>Value</span>
             <div className="group relative flex justify-center">
               <button>
@@ -49,7 +49,7 @@ export default async function BasicInfo({ player }) {
               </span>
             </div>
           </dt>
-          <dd className="text-sm sm:text-base leading-6 text-slate-700 dark:text-slate-200 text-right sm:text-left col-span-2 capitalize flex items-center justify-end sm:justify-start">
+          <dd className="text-sm @[16rem]/inner:text-base leading-6 text-slate-700 dark:text-slate-200 text-right @[16rem]/inner:text-left col-span-2 capitalize flex items-center justify-end @[16rem]/inner:justify-start">
             <Suspense
               fallback={
                 <SpinnerIcon className="animate-spin h-4 w-4 text-slate-400" />
