@@ -1,7 +1,9 @@
 import Image from "next/image";
 import Link from "next/link";
-import { ArrowTopRightOnSquareIcon } from "@heroicons/react/20/solid";
-import { ToggleFavouriteButton } from "../Favourites/ToggleFavouriteButton";
+import {
+  ArrowTopRightOnSquareIcon,
+  ArrowsRightLeftIcon,
+} from "@heroicons/react/20/solid";
 import SpinnerIcon from "../SpinnerIcon";
 import { Suspense } from "react";
 import { FavouriteButton } from "./FavouriteButton";
@@ -16,9 +18,9 @@ function LoadingFavouriteButton() {
 
 export default function ImageCard({ playerId }) {
   return (
-    <div className="mx-auto max-w-xs sm:mx-0">
+    <div className="mx-auto max-w-xs @sm:mx-0">
       <Image
-        className="-mt-2"
+        className="-mt-2 max-w-[200px] w-full mx-auto @sm:max-w-none"
         src={`https://d13e14gtps4iwl.cloudfront.net/players/${playerId}/card_512.png`}
         alt={`Player ${playerId}`}
         width="512"
@@ -30,6 +32,18 @@ export default function ImageCard({ playerId }) {
         <Suspense fallback={<LoadingFavouriteButton />}>
           <FavouriteButton playerId={playerId} />
         </Suspense>
+        <Link
+          href={{
+            pathname: "/compare",
+            query: {
+              player1: playerId,
+              player2: "",
+            },
+          }}
+          className="flex items-center justify-center space-x-1.5 text-sm font-semibold bg-slate-100 hover:bg-slate-200 dark:bg-slate-800 dark:hover:bg-slate-800/50 px-3.5 py-2 rounded-lg cursor-pointer ring-1 ring-slate-950 dark:ring-slate-800 ring-opacity-5"
+        >
+          <ArrowsRightLeftIcon className="w-5 h-5 -mr-1" />
+        </Link>
         <Link
           href={`https://app.playmfl.com/players/${playerId}`}
           target="_blank"
