@@ -1,30 +1,34 @@
 export const attributeWeighting = [
   {
-    positions: ["ST"],
-    weighting: [0.1, 0.46, 0, 0.29, 0.1, 0.05],
-  },
-  {
-    positions: ["CF", "LW", "RW"],
-    weighting: [0.24, 0.23, 0, 0.4, 0.13, 0],
-  },
-  {
-    positions: ["CAM"],
-    weighting: [0.34, 0.21, 0, 0.38, 0.07, 0],
-  },
-  {
-    positions: ["CM", "LM", "RM"],
-    weighting: [0.43, 0.12, 0.1, 0.29, 0, 0.06],
-  },
-  {
-    positions: ["CDM"],
-    weighting: [0.28, 0, 0.4, 0.17, 0, 0.15],
-  },
-  {
-    positions: ["LWB", "RWB", "LB", "RB"],
-    weighting: [0.19, 0, 0.44, 0.17, 0.1, 0.1],
+    positions: ["GK"],
+    weighting: [0, 0, 0, 0, 0, 0, 1],
   },
   {
     positions: ["CB"],
-    weighting: [0.05, 0, 0.64, 0.09, 0.02, 0.2],
+    weighting: [0.05, 0, 0.64, 0.09, 0.02, 0.2, 0],
+  },
+  {
+    positions: ["LWB", "RWB", "LB", "RB"],
+    weighting: [0.19, 0, 0.44, 0.17, 0.1, 0.1, 0],
+  },
+  {
+    positions: ["CDM"],
+    weighting: [0.28, 0, 0.4, 0.17, 0, 0.15, 0],
+  },
+  {
+    positions: ["CM", "LM", "RM"],
+    weighting: [0.43, 0.12, 0.1, 0.29, 0, 0.06, 0],
+  },
+  {
+    positions: ["CAM"],
+    weighting: [0.34, 0.21, 0, 0.38, 0.07, 0, 0],
+  },
+  {
+    positions: ["CF", "LW", "RW"],
+    weighting: [0.24, 0.23, 0, 0.4, 0.13, 0, 0],
+  },
+  {
+    positions: ["ST"],
+    weighting: [0.1, 0.46, 0, 0.29, 0.1, 0.05, 0],
   },
 ];
